moved mouse logic in mouse class
